ml-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
firebase-admin==6.1.0
python-dotenv==1.0.0
pydantic==2.5.0
joblib==1.3.2